update check using querySelector and confirmed working in console
in Javascript, it seems you cannot just pass the empty array as a falsy value - instead, to check whether a JS array is empty, you use .length. Anyway, I decided not go that way anyway and just go with querySelector. I used the img alt-text selector because it's more self-explanatory of what I'm pointing to in the DOM. I also added some comments for clarity