fix: if user clicks an operator without starting with a number, the operator should not show up
i know that i will eventually need to fix this to allow for negative number calculations. But, for now, the code has been fixed because the gate condition was previously wrong - i didn't check for smallDisplay.textContent, but just smallDisplay. which is a mistake, because smallDisplay is a DOM element that is always truthy. silly (beginner) mistake!