add backspace button instead of % calculations (#1)
* add backspace button and colour

i tried to use fill='currentColor'
to update the colour of the backspace vector image programmatically in css, but it didn't work, so i had to update it manually in the svg itself

* feat: add backspace button to delete a character that has been inputted by user

if the display is 0, nothing happens. if the display only has one character, then show 0, otherwise chop off the last digit. also, if the display is showing something that has been calculated, then do nothing also