feat: enable key digit input, and reduce max characters displayed
the code in the function displayDigitOrDecimalWithKeyboard is unclean. if the input is a digit, the arg passed into displayDigit function (which is common event handler, fired when user clicks the digits with their mouse) is event e, rather than the digit itself. ideally, and perhaps for future refactoring, displayDigit's parameter is the actual digit input, which can either be the e.key of a keyboard event, or e.target.textContent of the button element being clicked