feat: enable using keyboard (#2)
* add todo and placeholder function

* test alerting the .key attribute of event e parameter

* enable decimal point to be added via keyboard

* fix: double decimals

changed logic around so that if the last character is already a decimal, the decimal pressed again just removes it

* update event type to keydown so that users can hold a key for multiple numbers

* feat: enable key digit input, and increase max characters displayed

the code in the function displayDigitOrDecimalWithKeyboard is unclean. if the input is a digit, the arg passed into displayDigit function (which is common event handler, fired when user clicks the digits with their mouse) is event e, rather than the digit itself. ideally, and perhaps for future refactoring, displayDigit's parameter is the actual digit input, which can either be the e.key of a keyboard event, or e.target.textContent of the button element being clicked

* feat: enable key digit input, and reduce max characters displayed

the code in the function displayDigitOrDecimalWithKeyboard is unclean. if the input is a digit, the arg passed into displayDigit function (which is common event handler, fired when user clicks the digits with their mouse) is event e, rather than the digit itself. ideally, and perhaps for future refactoring, displayDigit's parameter is the actual digit input, which can either be the e.key of a keyboard event, or e.target.textContent of the button element being clicked

* use dispatch event to force a mousedown and mouseup, then click

* programmatically add the digits as a class for each digit button

this makes it easier to identify when mapping keyboard .key's to the buttons

* ensure the class that i am adding to each of the digits don't start with a number

this is not allowed in css

* make decimal button change colour once selected and held

* use else if statement as clearer

* click decimal button and equivalent digit button, identified via class

I believe, right now as I'm writing, that this code is more readable by routing all key strokes to the equivalent button on the calculator, rather than calling the functions that are triggered by mouseclicks, directly

* align event listeners on the entire page

* feat: clicking digit buttons inputs into display, and also changes the style of the button (looks like it's been pressed)

* fix: backspace button was on hold when it was meant to come off clicked once you let go of mouse on the button

* rename function to be more inclusive of other buttons

* add underscores to all capitalised constants

* fix: update variable references to max_chars_displayed (with underscores). reduce indentation by using return statements in conditionals

* start creating hashmap where key is the button, and values are the keys that would trigger them

had to do it this way (rather than the key of object being keyboard key) as there could be multiple keys to trigger a button, but you need to have unique keys in an object

* add hashmap of operators, where keys are the operator (shown in html), and values are the keyboard keys

* find equivalent button by looking through all button's textContent

* create set of operator keyboard keys

use set instead of array because faster. i think this is best practice. checking whether the key is inside of it (coming up next commit) is also more intuitive imo with .has method

* feat: enable keyboard input for operators (+, -, *times, divide)

* using hashmap, find the button to click if an operator button is clicked

i tried to use || to find the right button, but it kept defaulting to divide if i used +. this means i had to hardcode in that '+' means '+' which feels really dumb. I also have two ways (used to be 3, if i included capital X, which i've now removed by lowercasing everything at the start) to map to multiplication, which again feels silly, but it's way easier to hardcode this at my skill level right now, from what i can find on google, to find the equivalent _key_ of an object, if a value is in the _value_ (as array) of an object

* feat: backspace on keyboard (or delete) does backspace/delete on display

* enable equals button to be pressed via enter key or equals key on keyboard. also ensured it isn't 'stuck'

* click escape button with keyboard

* move variables to top of function for clarity

* write ternary statement to get equivalentBtn

* use ternary statements to conditionally assign equivalent button (to key), which is then mousedown'd (to change its colour) and then clicked (to trigger its functionality)

i had to put in an exit condition because using ternary statements was the cleanest way, in my opinion. otherwise, using normal conditionals, i'd have to type out 'equivalentBtn' variables multiple times.

* fix: esc did not refresh page